pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sandgnat-orchestrator"
version = "0.1.0"
description = "Malware runtime analysis job orchestrator (Celery, STIX 2.1, Postgres)"
requires-python = ">=3.11"
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "celery>=5.3",
    "redis>=5.0",
    "psycopg[binary,pool]>=3.1",
    "stix2>=3.0",
    "proxmoxer>=2.0",
    "requests>=2.31",
    "python-dateutil>=2.9",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
pcap = [
    # Optional: only needed on hosts that parse PCAPs.
    "scapy>=2.5",
]
yara = [
    # Optional: enable pre-detonation YARA scanning in the intake pipeline.
    "yara-python>=4.5",
]
static = [
    # Linux static-analysis guest dependencies. Install on the Linux
    # static-analysis VM (not the orchestrator host). All tools are optional
    # at runtime — missing libs degrade to a no-op rather than crashing.
    "pefile>=2023.2.7",
    "pyelftools>=0.30",
    "ssdeep>=3.4",
    "python-tlsh>=4.5",
    "capstone>=5.0",
]

[project.scripts]
sandgnat-worker = "orchestrator.worker:main"
sandgnat-intake = "orchestrator.intake_server:main"

[tool.setuptools.packages.find]
include = ["orchestrator*", "linux_guest_agent*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
